+++ build_test/test/test_builder_test.dart
@@ -169,13 +169,14 @@ void main() {
           '.dart': ['.fake']
         },
         build: (step, _) async {
-          var lib = await step.inputLibrary;
-          var errors = await lib.session.getErrors(lib.source.fullName);
           expect(
-              errors.errors.map((e) => e.message),
-              contains(contains(
-                  'This requires the \'extension-methods\' language feature to be '
-                  'enabled.')));
+              step.inputLibrary,
+              throwsA(isA<SyntaxErrorInAssetException>().having(
+                  (e) => e.syntaxErrors.first.message,
+                  'message',
+                  contains(
+                      'This requires the \'extension-methods\' language feature to be '
+                      'enabled.'))));
         },
       ),
       {